NVDynamic: Fix use of uninitialized value.
There is a theoretical possibility that the cycle in NvNextByType does
not run at all because NvNext fails immediately and never sets nvHandle.

The cycle only runs and nvHandle is set on non-zero addr, and in that
case addr remains non-zero when breaking out of the cycle.

Use nvHandle only when addr is non-zero.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>