Fix rev1332 RSA Primary problem
The routine RsaAdjustPrimeCandidate changed after rev1332 in such a
way that it produces a different RSA primary key from the same seed.
In order not to have the RSA primary change across upgrade, we have to
go back to the original method for RsaAdjustPrimeCandidate if the
firmware is originally from rev1332.

Steal the last four bytes of the NVChip file to represent a firmware
version.  Set it to 0 if the firmware came from rev1332 and 1
otherwise.  If a 0 is seen here, we can now invoke the older
RsaAdjustPrimeCandidate function so the RSA primary seed does not
change.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>