tpm2: Initialize a whole OBJECT before using it
Copied from https://github.com/stefanberger/libtpms/blob/stable-0.9/src/tpm2/Object.c#L280
libtpms date: Thu, 22 Jul 2021 21:23:58 -0400

Initialize a while OBJECT before using it. This is necessary since
an OBJECT may also be used as a HASH_OBJECT via the ANY_OBJECT
union and that HASH_OBJECT can leave bad size inidicators in TPM2B
buffer in the OBJECT. To get rid of this problem we reset the whole
OBJECT to 0 before using it. This is as if the memory for the
OBJECT was just initialized.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>
Signed-off-by: Ken Goldman <kgold@linux.ibm.com>