tpm: Object: clear out sensitive area if on load if not provided
libtpms commit of Sat, 5 Aug 2023 11:44:23 +0100

When an object is later marshalled in TPM2_ContextSave, the publicOnly
attribute isn't taken into account and therefore potentially stale
sensitive information can be marshalled, which is a problem if the
buffer sizes it contains have values that are too large - this
triggers assertion failures.

Avoid this by clearing out the sensitive area upon ObjectLoad if not
provided, making the behaviour consistent with when a fresh, unused,
object entry is used.

Signed-off-by: Rob Shearman <rob@graphiant.com>
Signed-off-by: Ken Goldman <kgold@linux.ibm.com>